Fix how book-appointment gets identity data
This fixes a bug where the frontend will display "early" (<5d) appointments
to logged-in (admin) users but a subsequent call to book/appointment will
throw a :window-unavailable error. Refactor to pass user as a separate param
to both book-appointment and get-available-windows, to decrease the chance
of accidentally introducing a vulnerability later. Then, pass user (identity)
as second arg to both book-appointment and get-available-windows